requirements.txt
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
decouple==0.0.7
dj-database-url==2.3.0
Django==5.0.7
django-environ==0.11.2
django-phonenumber-field==8.0.0
django-storages==1.14.4
environ==1.0
google-api-core==2.23.0
google-auth==2.36.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
gunicorn==23.0.0
idna==3.10
packaging==24.2
phonenumbers==8.13.48
proto-plus==1.25.0
protobuf==5.28.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyMySQL==1.1.1
pyngrok==7.2.0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
rsa==4.9
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
whitenoise==6.8.2
